# server/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DD Core API",
    "version": "0.1.0",
    "description": "Stable v1 contract for DD instance discovery, service health, account authentication, contacts, and P4 reliable text messaging/sync. P0 call endpoints remain experimental and are intentionally excluded until promoted into /api/v1.",
    "license": {
      "name": "License decision pending",
      "url": "https://dd.invalid/license-pending"
    }
  },
  "servers": [
    {
      "url": "https://{instanceHost}",
      "description": "Self-hosted DD instance",
      "variables": {
        "instanceHost": {
          "default": "dd.invalid",
          "description": "Hostname of the self-hosted instance"
        }
      }
    }
  ],
  "security": [],
...
  "paths": {
...
    "/api/v1/media/uploads": {
      "post": {
        "operationId": "createMediaUpload",
        "summary": "Reserve a private media upload",
        "tags": ["Media"],
        "security": [{"bearerAuth": []}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateMediaUploadRequest"}}}},
        "responses": {
          "201": {"description": "Short-lived object-storage upload grant", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MediaUploadGrantResponse"}}}},
          "400": {"description": "Invalid media metadata", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "413": {"description": "Upload exceeds quota or size policy", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "503": {"description": "Media service unavailable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
...
    "/api/v1/media/{mediaId}": {
      "get": {
        "operationId": "getMediaMetadata",
        "summary": "Get authorized media metadata",
        "tags": ["Media"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/MediaId"}],
        "responses": {
          "200": {"description": "Media metadata", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MediaObjectResponse"}}}},
          "401": {"description": "Access token is missing or invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "403": {"description": "Current user cannot access this media", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}},
          "404": {"description": "Media not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}}
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "MediaId": {"name": "mediaId", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreateMediaUploadRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["fileName", "size", "mimeType", "sha256", "purpose"],
        "properties": {
          "fileName": {"type": "string", "minLength": 1, "maxLength": 255},
          "size": {"type": "integer", "format": "int64", "minimum": 1},
          "mimeType": {"type": "string", "minLength": 1, "maxLength": 120},
          "sha256": {"type": "string", "pattern": "^[a-fA-F0-9]{64}$"},
          "purpose": {"type": "string", "enum": ["CHAT_IMAGE", "CHAT_FILE", "CHAT_VOICE", "CHAT_VIDEO", "STICKER", "GIF", "MOMENT_IMAGE", "MOMENT_VIDEO", "GROUP_AVATAR", "MOMENT_COVER"]}
        }
      },
      "MediaUploadGrant": {
        "type": "object",
        "additionalProperties": false,
        "required": ["uploadId", "mediaId", "uploadUrl", "expiresAt", "requiredHeaders"],
        "properties": {
          "uploadId": {"type": "string", "format": "uuid"},
          "mediaId": {"type": "string", "format": "uuid"},
          "uploadUrl": {"type": "string", "format": "uri"},
          "expiresAt": {"type": "string", "format": "date-time"},
          "requiredHeaders": {"type": "object", "additionalProperties": {"type": "string"}}
        }
      },
      "MediaUploadGrantResponse": {
        "type": "object", "additionalProperties": false, "required": ["data", "requestId"],
        "properties": {"data": {"$ref": "#/components/schemas/MediaUploadGrant"}, "requestId": {"$ref": "#/components/schemas/RequestIdValue"}}
      },
      "MediaObject": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "originalName", "mimeType", "sizeBytes", "sha256", "purpose", "status", "encryptionMode", "createdAt"],
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "ownerUserId": {"type": "string", "format": "uuid"},
          "originalName": {"type": "string"},
          "mimeType": {"type": "string"},
          "sizeBytes": {"type": "integer", "format": "int64", "minimum": 0},
          "sha256": {"type": "string"},
          "purpose": {"type": "string", "enum": ["CHAT_IMAGE", "CHAT_FILE", "CHAT_VOICE", "CHAT_VIDEO", "STICKER", "GIF", "MOMENT_IMAGE", "MOMENT_VIDEO", "GROUP_AVATAR", "MOMENT_COVER"]},
          "status": {"type": "string", "enum": ["UPLOADING", "READY", "QUARANTINED", "FAILED", "DELETED"]},
          "encryptionMode": {"type": "string"},
          "createdAt": {"type": "string", "format": "date-time"},
          "readyAt": {"type": "string", "format": "date-time"}
        }
      },
      "MediaObjectResponse": {
        "type": "object", "additionalProperties": false, "required": ["data", "requestId"],
        "properties": {"data": {"$ref": "#/components/schemas/MediaObject"}, "requestId": {"$ref": "#/components/schemas/RequestIdValue"}}
      },
...
      "RequestIdValue": {
        "type": "string",
        "pattern": "^req_[A-Za-z0-9]+$"
      },
      "APIError": {
        "type": "object",
        "additionalProperties": true,
        "required": ["code", "message", "requestId"],
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]*$"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "$ref": "#/components/schemas/RequestIdValue"
          },
          "details": {}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/APIError"
          }
        }
      }
    }
  }
}